backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
drf-yasg>=1.21.0
Pillow>=10.0.0
celery>=5.3.0
redis>=4.5.0
